Add error handling for empty choices
When the `Tty::Prompt#select` method is called with an empty list of
choices it fails hard causing a stacktrace to be printed. This is fine
in theory but means that we need to check if the choice list is empty
before doing that.

Error handling for this case has been added to `RfcReader::Command`.

It also made sense to exit with a non-zero error code in those cases
but using `Kernel.exit` or `Kernel.abort` here is a bit problematic
since it causes the rspec to fail while testing those failures cases.

The compromise is to return an integer exit code from all commands
in `RfcReader::Command` and then just exit with that error code
in `exe/rfc-reader`. This is a bit ugly but seems to work fine.

Rspec has also been cleaned up a bit and new tests were added to
cover the most obvious failure cases.