Add ErrorContext helper
This allows you to add context to regions of code by wrapping them
in blocks with strings as context. If an error is raised in the
block, it gets wrapped with the context and re-raised.

There is also a handler which should wrap the majority of the app
logic that handles how these errors are displayed to users. By
default it doesn't make sense to print stacktraces for normal users
even with a command line application so we only print the short
message and a note about setting the `DEBUG` environment variable
to see the full message.